@@ -24,36 +24,36 @@
     "publish:mac": "vue-cli-service electron:build --mac -p always"
   },
   "dependencies": {
-    "vue": "^2.5.17",
-    "vue-router": "^3.0.1",
+    "vue": "^2.5.21",
+    "vue-router": "^3.0.2",
     "vuex": "^3.0.1"
   },
   "devDependencies": {
     "@trodi/electron-splashscreen": "^0.3.4",
-    "@vue/cli-plugin-babel": "^3.1.1",
-    "@vue/cli-plugin-eslint": "^3.1.5",
-    "@vue/cli-service": "^3.2.2",
+    "@vue/cli-plugin-babel": "^3.3.0",
+    "@vue/cli-plugin-eslint": "^3.3.0",
+    "@vue/cli-service": "^3.3.0",
     "babel-eslint": "^10.0.1",
     "electron": "^3.0.0",
     "electron-log": "^2.2.17",
     "electron-updater": "^4.0.6",
-    "eslint": "^5.8.0",
-    "eslint-plugin-vue": "^5.0.0-0",
+    "eslint": "^5.12.0",
+    "eslint-plugin-vue": "^5.1.0",
     "fs-extra": "^7.0.1",
-    "less": "^3.0.4",
+    "less": "^3.9.0",
     "less-loader": "^4.1.0",
     "lodash": "^4.17.11",
-    "moment": "^2.22.2",
+    "moment": "^2.23.0",
     "progress-stream": "^2.0.0",
     "promise-ftp": "^1.3.5",
     "request": "^2.88.0",
     "request-promise": "^4.2.2",
     "rsync": "^0.6.1",
-    "scrape-it": "^5.1.1",
+    "scrape-it": "^5.1.4",
     "transmission-client": "^1.0.0",
-    "uikit": "^3.0.0-rc.23",
-    "vue-cli-plugin-electron-builder": "^1.0.0-rc.8",
-    "vue-template-compiler": "^2.5.17"
+    "uikit": "^3.0.0-rc.26",
+    "vue-cli-plugin-electron-builder": "^1.0.0-rc.10",
+    "vue-template-compiler": "^2.5.21"
   },
   "main": "background.js"
 }
